+++ server/src/main/java/consamables/api/MenuSection.java
@@ -12,7 +12,7 @@ import consamables.jdbi.ItemDAO;
 public class MenuSection {
     @NotNull
     @JsonProperty
-    private int menuSectionId;
+    private long menuSectionId;
     
     @NotNull
     @JsonProperty
@@ -22,17 +22,17 @@ public class MenuSection {
     @JsonProperty
     private List<Item> items;
 
-    public MenuSection(int menuSectionId, String name) {
+    public MenuSection(long menuSectionId, String name) {
         this.menuSectionId = menuSectionId;
         this.name = name;
         this.items = new ArrayList<Item>();
     }
 
-    public int getMenuSectionId() {
+    public long getMenuSectionId() {
         return menuSectionId;
     }
 
-    public void setMenuSectionId(int menuSectionId) {
+    public void setMenuSectionId(long menuSectionId) {
         this.menuSectionId = menuSectionId;
     }
