+++ server/src/main/java/consamables/api/Restaurant.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 public class Restaurant {
 
     @JsonProperty
-    private Integer restaurantId;
+    private Long restaurantId;
     
     @NotNull
     @JsonProperty
@@ -28,7 +28,7 @@ public class Restaurant {
     
     public Restaurant() { }
 
-    public Restaurant(Integer restaurantId, String name, JsonNode location, JsonNode hours, String url) {
+    public Restaurant(Long restaurantId, String name, JsonNode location, JsonNode hours, String url) {
         this(name, location, hours, url);
         this.restaurantId = restaurantId;
     }
@@ -40,11 +40,11 @@ public class Restaurant {
         this.url = url;
     }
 
-    public Integer getRestaurantId() {
+    public Long getRestaurantId() {
         return restaurantId;
     }
 
-    public void setRestaurantId(Integer restaurantId) {
+    public void setRestaurantId(Long restaurantId) {
         this.restaurantId = restaurantId;
     }
